refactor(rust): move shared HTTP/CLI transport helpers into src/transport
Byte-identical moves from providers/mod.rs; old import paths preserved via
pub(crate) re-exports, so provider files carry zero import churn.
