+++ sdks/rust/src/client.rs
@@ -3,33 +3,9 @@ use crate::providers::Provider;
 use crate::retry::RetryPolicy;
 use crate::stream::BoxStream;
 use crate::think_stripper::ThinkStripper;
+use crate::transport::{TimeoutConfig, DEFAULT_CONNECT_TIMEOUT, DEFAULT_READ_IDLE_TIMEOUT};
 use crate::types::{ChatRequest, ChatResponse, Message, StreamEvent, StreamEventType};
 use std::time::Duration;
-
-pub(crate) const DEFAULT_CONNECT_TIMEOUT: Duration = Duration::from_secs(10);
-pub(crate) const DEFAULT_READ_IDLE_TIMEOUT: Duration = Duration::from_secs(120);
-
-/// Unified timeout model:
-/// - `connect`: TCP/TLS connect deadline on the shared reqwest client.
-/// - `read_idle`: max gap between HTTP stream chunks before
-///   `MotosanError::StreamReadTimeout`.
-/// - `total`: opt-in wall-clock budget per blocking `chat()` attempt.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) struct TimeoutConfig {
-    pub(crate) connect: Duration,
-    pub(crate) read_idle: Duration,
-    pub(crate) total: Option<Duration>,
-}
-
-impl Default for TimeoutConfig {
-    fn default() -> Self {
-        Self {
-            connect: DEFAULT_CONNECT_TIMEOUT,
-            read_idle: DEFAULT_READ_IDLE_TIMEOUT,
-            total: None,
-        }
-    }
-}
 
 /// The provider selected and fully constructed at `ClientBuilder::build()`
 /// time. Dispatch borrows this instead of rebuilding per request.
